refactoring - common user download from HTTP
